+++ src/agents/__init__.py
@@ -1,3 +1,3 @@
-"""A Python Project."""
+"""Build Agents with tool use."""
 
 __version__ = "0.0.1"
